+++ src/index.html
@@ -8,7 +8,7 @@
     <!-- Link to the Font-Awesome Font -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.2/css/font-awesome.min.css" />
     <!-- CSS only -->
-    <link rel="stylesheet" href="style.css" type="text/css" />
+    <link rel="stylesheet" type="text/css" href="./style.css" />
 </head>
 
 <body>
@@ -24,6 +24,6 @@
         </form>
     </section>
     <script type="module" src="index.js"></script>
-<script defer src="main.js"></script></body>
+</body>
 
 </html>
