+++ src/module/constractors.js
@@ -1,0 +1,7 @@
+export default class MyObject {
+  constructor(description, completed, index) {
+    this.description = description;
+    this.completed = completed;
+    this.index = index;
+  }
+}
